Reimplement automata for better tokenization and errors
It was not clear how to create tokens with previous
implementation. Now SPACE states have been substituted
by TENTATIVE_TOKEN and TOKEN_END sates. These help
control when a string needs to be tokenized for sure or could
continue without being finished.

Also added error states when ended the evaluation.
In a matrix of strs have filled the ones that should
produce an error state with its custom message.