Add tokenize into binary tree feature
To sort all the commands a binary tree is implemented.
The automata recognizes if the read character
is part of a command or if its and operator and
inserts the token into the binary tree creating
the following structure:
			op1
	comm1				op2
	args1			comm2		comm3
				args2		args3

So all nodes to the right are operators that have
on the left their left command and on the right
their right command (nested into operators).