Merge new double automata into previous code
Now we have two automatas, lexer and parser.
To make everything work smoothly I have separated
them, risking a bit of repetition. Now there is a
lexer for the first sweep and a parser to read
each specific command and this will execute the
expander. Also detached completely the AOE from
the automata, so we just give a reference to it
and "pray" that it was correctly initialized.